feat: new features and changes

+++ src/Dashboard/Profile/Experience/Education/FetchEducation/DeleteEducation.js
@@ -1,7 +1,12 @@
 const DeleteEducation = async (email)=>{
-    return await fetch(`http://localhost:8080/api/user-education/delete/${email}`, {
-        method: "DELETE",
-    });
-
+    const token = localStorage.getItem("token")
+    if(token) {
+        return await fetch(`http://localhost:8080/api/user-education/delete/${email}`, {
+            method: "DELETE",
+            headers: {
+                'Authorization': `Bearer ${token}`
+            }
+        });
+    }
 }
 export default DeleteEducation;
