+++ src/Dashboard/Profile/Experience/Work Experience/FetchWorkExperience/DeleteWorkExperience.js
@@ -1,6 +1,12 @@
 const DeleteWorkExperience = async (email)=>{
-    return await fetch(`http://localhost:8080/api/work-experience/delete/${email}`, {
-        method: "DELETE",
-    });
+    const token = localStorage.getItem("token")
+    if(token) {
+        return await fetch(`http://localhost:8080/api/work-experience/delete/${email}`, {
+            method: "DELETE",
+            headers: {
+                'Authorization': `Bearer ${token}`
+            }
+        });
+    }
 }
 export default DeleteWorkExperience
